LPD-47216 Separate instance/site Export-Import functional tests
